Fix $setCellStateForNextGeneration with 8 living neighbors
Your lookup table was designed to handle 0 to 7 living neighbors. For dead cells with 8 living neighbors the result is correct because it overflows into the section of the table that's meant for living cells. For a living cell with 8 living neighbors you would get an index of 16 which is out of bounds.